% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_default.R
\name{gt_default}
\alias{gt_default}
\title{GT Default Display}
\usage{
gt_default(df, custom_caption = NULL)
}
\arguments{
\item{df}{a data frame}

\item{custom_caption}{a table title}
}
\description{
GT Default Display
}
