% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_default.R
\name{gtsum_default}
\alias{gtsum_default}
\title{GT Default Display for Table Summary/Regression Obj}
\usage{
gtsum_default(tbl_obj, custom_caption = NULL)
}
\arguments{
\item{tbl_obj}{a tbl_summary or tbl_regression object}
}
\description{
GT Default Display for Table Summary/Regression Obj
}
