% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwiktools-package.R
\docType{package}
\name{kwiktools-package}
\alias{kwiktools}
\alias{kwiktools-package}
\title{kwiktools: Small Utilities for Everyday Data Analysis Tasks}
\description{
A collection of small utility functions for common data analysis tasks and workflows.
}
\author{
\strong{Maintainer}: Karissa Whiting \email{karissa.whiting@gmail.com} (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})

}
\keyword{internal}
