% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_to_labels.R
\name{names_to_labels}
\alias{names_to_labels}
\title{Convert DF names to labels}
\usage{
names_to_labels(df, caps = NULL)
}
\arguments{
\item{df}{a data frame}

\item{caps}{a string of capitalization exceptions}
}
\description{
Convert DF names to labels
}
