% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-date.R
\name{save_with_date}
\alias{save_with_date}
\title{Wrapper for \code{save()} and \code{readr::write_csv()} that
saves a RData or csv with today's date appended to filename}
\usage{
save_with_date(object, path)
}
\arguments{
\item{object}{the name of the object to be saved}

\item{path}{the path of the data as you would normally provide it to \code{save()}
or \code{readr::write_csv()}}
}
\description{
Wrapper for \code{save()} and \code{readr::write_csv()} that
saves a RData or csv with today's date appended to filename
}
\examples{
\dontrun{
save_with_date(df, here::here("data", "patients.RData"))
}
}
