% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwiktools-package.R
\docType{package}
\name{kwiktools-package}
\alias{kwiktools}
\alias{kwiktools-package}
\title{kwiktools: function dumpster}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
