% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-most-recent.R
\name{load_most_recent}
\alias{load_most_recent}
\title{Wrapper for \code{load()} and \code{readr::read_csv()} that
saves a RData or csv with today's date appended to filename}
\usage{
load_most_recent(directory, filename_keyword = NULL, date_in_filename = TRUE)
}
\arguments{
\item{directory}{directory to look for most recent data files in}

\item{filename_keyword}{a file name (e.g. "data.RData") or
keyword (e.g. "patient-data"). If provided, function will only consider files
in that directory with that name or keyword}

\item{date_in_filename}{If \code{TRUE}, function will use dates in filenames to sort.
If \code{FALSE} function will use the date the file was last updated to sort.
\code{TRUE} by default.}
}
\description{
Wrapper for \code{load()} and \code{readr::read_csv()} that
saves a RData or csv with today's date appended to filename
}
\examples{
save_date(df, here::here("data", "patients.RData"))
}
