% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_to_labs.R
\name{names_to_labs}
\alias{names_to_labs}
\title{Convert DF names to labels}
\usage{
names_to_labs(names_df)
}
\arguments{
\item{df}{a data frame}

\item{caps}{a string of capitalization exceptions}
}
\description{
Convert DF names to labels
}
